refactor(users, posts): Refactorización de como deberia quedar separado por modulos
