+++ src/posts/posts.module.ts
@@ -1,0 +1,12 @@
+
+import { Module } from '@nestjs/common';
+import { PostsController } from './posts.controller';
+import { PostsService } from './posts.service';
+import { PrismaModule } from '../prisma/prisma.module';
+
+@Module({
+  imports: [PrismaModule],
+  controllers: [PostsController],
+  providers: [PostsService],
+})
+export class PostsModule {}
