+++ src/posts/posts.service.ts
@@ -1,7 +1,7 @@
 
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
-import { Post, Prisma } from './generated/prisma/client';
+import { PrismaService } from '../prisma/prisma.service';
+import { Post, Prisma } from '../generated/prisma/client';
 
 @Injectable()
 export class PostsService {
